3.1-drf-intro/smart_home/requirements.txt
django~=4.0.3
djangorestframework~=3.13.1
psycopg2-binary
